fixed mulitiple bugs in trajectory cacluations
but there is still a bug remaining at numer 49 in create_sin_traj. Only if i negated the skew(rot_ax) the sign of omega_d was correct. It seems that in the function find_rotation_axis is an error. If you set in the values of rot_ax and rot_alpha_scale into the rodriguesz formula it doesn't result in R_target!! https://github.com/Emucde/MA24_simulink/issues/80